move files and update license
